requirements.txt
Django==5.2.1
djangorestframework
djangorestframework-simplejwt

